pyproject.toml
[tool.poetry]
name = "raggy"
version = "0.1.0"
description = "Raggy - An AI exploration system using Iroh and Streamlit"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.29.0"
iroh = "0.27.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
huggingface-hub = "^0.19.4"
tokenizers = "^0.15.0"
pillow = "^10.1.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


